backend/requirements.txt
fastapi
uvicorn
pandas
python-multipart
scipy
scikit-learn
numpy
xgboost
joblib
pyarrow
fastavro
avro-python3